% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.chi_cal.R
\name{r.chi_cal}
\alias{r.chi_cal}
\title{Reduced chi-square of a model}
\usage{
r.chi_cal(y, model)
}
\arguments{
\item{y}{A list of numeric value the response in the model, can be a column of an existing tibble.}

\item{model}{A fitted model either by linear fitting or non-linear fitting.}
}
\value{
A single numeric
}
\description{
Get a number of reduced chi-square.
}
